+++ src-tauri/src/commands/osc.rs
@@ -1,22 +1,82 @@
 use std::{
     net::{Ipv4Addr, SocketAddrV4, UdpSocket},
     str::FromStr,
+    sync::{mpsc, Mutex},
 };
 
 use log::{debug, error};
 use rosc::{encoder, OscMessage, OscPacket, OscType};
+use serde::{Deserialize, Serialize};
 
-use crate::OSC_SOCKET;
+use crate::{background, OSC_RECEIVE_SOCKET, OSC_SEND_SOCKET};
+
+#[derive(Clone, Serialize, Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct OSCValue {
+    pub kind: String,
+    pub value: Option<String>,
+}
+#[derive(Clone, Serialize, Deserialize, Debug)]
+#[serde(rename_all = "camelCase")]
+pub struct OSCMessage {
+    pub address: String,
+    pub values: Vec<OSCValue>,
+}
+
+lazy_static! {
+    static ref TERMINATION_TX: Mutex<Option<mpsc::Sender<()>>> = Default::default();
+}
 
 #[tauri::command]
-pub fn osc_init() -> bool {
-    *OSC_SOCKET.lock().unwrap() = match UdpSocket::bind((Ipv4Addr::UNSPECIFIED, 0)) {
+pub fn osc_init(receive_addr: String) -> bool {
+    // Terminate existing thread if it exists
+    let termination_guard = TERMINATION_TX.lock().unwrap();
+    let termination = termination_guard.as_ref();
+    if termination.is_some() {
+        termination.unwrap().send(()).unwrap();
+    }
+    drop(termination_guard);
+    // Setup sending socket
+    *OSC_SEND_SOCKET.lock().unwrap() = match UdpSocket::bind((Ipv4Addr::UNSPECIFIED, 0)) {
         Ok(s) => Some(s),
         Err(err) => {
-            error!("[Core] Could not initialize OSC module: {}", err);
+            error!(
+                "[Core] Could not initialize send socket for OSC module: {}",
+                err
+            );
             return false;
         }
     };
+    // Setup receiving socket if needed
+    let receive_socket_guard = OSC_RECEIVE_SOCKET.lock().unwrap();
+    if receive_socket_guard.is_none() {
+        drop(receive_socket_guard);
+        let receive_addr = match SocketAddrV4::from_str(receive_addr.as_str()) {
+            Ok(addr) => addr,
+            Err(err) => {
+                error!(
+                    "[Core] Could not initialize receive socket for OSC module (addr init): {}",
+                    err
+                );
+                return false;
+            }
+        };
+        let receive_socket = match UdpSocket::bind(receive_addr) {
+            Ok(s) => s,
+            Err(err) => {
+                error!(
+                    "[Core] Could not initialize receive socket for OSC module (socket init): {}",
+                    err
+                );
+                return false;
+            }
+        };
+        receive_socket.set_nonblocking(true).unwrap();
+        *OSC_RECEIVE_SOCKET.lock().unwrap() = Some(receive_socket);
+    }
+    // Process incoming messages
+    let termination_tx = background::osc::spawn_osc_receiver_thread();
+    *TERMINATION_TX.lock().unwrap() = Some(termination_tx);
     true
 }
 
@@ -57,7 +117,7 @@ pub fn osc_valid_addr(addr: String) -> bool {
 
 fn osc_send(addr: String, osc_addr: String, data: Vec<OscType>) -> Result<bool, String> {
     // Get socket
-    let socket_guard = OSC_SOCKET.lock().unwrap();
+    let socket_guard = OSC_SEND_SOCKET.lock().unwrap();
     let socket = match socket_guard.as_ref() {
         Some(socket) => socket,
         None => return Err(String::from("NO_SOCKET")),