+++ src/app/services/osc.service.ts
@@ -7,6 +7,9 @@ import { OscScript, OscScriptSleepAction } from '../models/osc-script';
 import { cloneDeep } from 'lodash';
 import { TaskQueue } from '../utils/task-queue';
 import { debug, info } from 'tauri-plugin-log-api';
+import { listen } from '@tauri-apps/api/event';
+import { OSCMessage, OSCMessageRaw, parseOSCMessage } from '../models/osc-message';
+import { Observable, Subject } from 'rxjs';
 
 @Injectable({
   providedIn: 'root',
@@ -14,11 +17,13 @@ import { debug, info } from 'tauri-plugin-log-api';
 export class OscService {
   address = '127.0.0.1:9000';
   private scriptQueue: TaskQueue = new TaskQueue({ runUniqueTasksConcurrently: true });
+  private _messages: Subject<OSCMessage> = new Subject<OSCMessage>();
+  public messages: Observable<OSCMessage> = this._messages.asObservable();
 
   constructor(private sleep: SleepService) {}
 
   async init() {
-    const result = await invoke<boolean>('osc_init');
+    const result = await invoke<boolean>('osc_init', { receiveAddr: '127.0.0.1:9001' });
     if (!result) {
       info(
         '[OSC] Could not bind a UDP socket to interact with VRChat over OSC (possibly due to incorrectly configured permissions). Quitting...'
@@ -28,7 +33,11 @@ export class OscService {
         { type: 'error', title: 'Oyasumi' }
       );
       await exit(0);
+      return;
     }
+    listen<OSCMessageRaw>('OSC_MESSAGE', (data) => {
+      this._messages.next(parseOSCMessage(data.payload));
+    });
   }
 
   async send_float(address: string, value: number) {